Fixing first file / last file logic

+++ skimmer/make_skims.sh
@@ -58,6 +58,7 @@ dataset_names=(
     TTJets_HT-2500toInf
 )
 
+
 make_skims() {
 
     local dataset_directory=$1
@@ -79,30 +80,34 @@ make_skims() {
         xrdfs ${output_redirector} mkdir -p ${output_dir}
     fi
 
+    i_file=-1
     for files_list in $(ls ${files_list_directory} | sort -V); do
+        ((i_file++))
+        if [ ${i_file} -le ${LAST_FILE} ]; then
+            if [ ${i_file} -ge ${FIRST_FILE} ]; then
 
-        local input_files=${files_list_directory}/${files_list}
-        local output_file=${output_directory}/${files_list/.txt/.root}
-        local output_file_name_tmp=$(echo ${ouput_file}_$(date +"%Y%m%d-%H%M%S") | shasum | cut -d " " -f1).root
-        local output_file_tmp=/uscmst1b_scratch/lpc1/3DayLifetime/${USER}/${output_file_name_tmp}
+                local input_files=${files_list_directory}/${files_list}
+                local output_file=${output_directory}/${files_list/.txt/.root}
+                local output_file_name_tmp=$(echo ${ouput_file}_$(date +"%Y%m%d-%H%M%S") | shasum | cut -d " " -f1).root
+                local output_file_tmp=/uscmst1b_scratch/lpc1/3DayLifetime/${USER}/${output_file_name_tmp}
 
-        echo ""
-        echo "Making skim file ${output_file}"
+                echo ""
+                echo "Making skim file ${output_file}"
 
-        local output_redirector=$(echo ${output_file} | cut -d/ -f 1-4)
-        local output_file_path=$(echo ${output_file} | cut -d/ -f 4-)
-        xrdfs ${output_redirector} ls ${output_file_path} > /dev/null 2>&1
-        if [ "$?" != "0" ] || [ "${FORCE_RECREATE}" == "1" ]; then
-            python skim.py -i ${input_files} -o ${output_file_tmp} -p ${module} -y ${year} -e ${EXECUTOR} -n ${N_WORKERS} -c ${CHUNK_SIZE} --memory ${MEMORY} --cores ${CORES}
-            xrdcp -f ${output_file_tmp} ${output_file}
-            echo ${output_file} has been saved.
-            rm ${output_file_tmp}
-        else
-            echo ${output_file} already exists and FORCE_RECREATE is 0. Skipping.
+                local output_redirector=$(echo ${output_file} | cut -d/ -f 1-4)
+                local output_file_path=$(echo ${output_file} | cut -d/ -f 4-)
+                xrdfs ${output_redirector} ls ${output_file_path} > /dev/null 2>&1
+                if [ "$?" != "0" ] || [ "${FORCE_RECREATE}" == "1" ]; then
+                    python skim.py -i ${input_files} -o ${output_file_tmp} -p ${module} -y ${year} -e ${EXECUTOR} -n ${N_WORKERS} -c ${CHUNK_SIZE} --memory ${MEMORY} --cores ${CORES}
+                    xrdcp -f ${output_file_tmp} ${output_file}
+                    echo ${output_file} has been saved.
+                    rm ${output_file_tmp}
+                else
+                    echo ${output_file} already exists and FORCE_RECREATE is 0. Skipping.
+                fi
+            fi
         fi
-        
     done
-
 }
 
 
@@ -111,3 +116,4 @@ for dataset_name in ${dataset_names[@]}; do
     make_skims ${dataset_directory} ${module} ${selection_name} ${year} ${dataset_name} ${output_directory}
 
 done
+
